server/pyproject.toml
[tool.black]
line-length = 100
target_version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.mypy_cache
    | \.pytest_cache
    | \.tox
    | \.venv
    | _build
    | build
    | dist
    | node_modules
    | public
    | src
    | staticfiles
  )/
)
'''
[tool.isort]
line_length = 100
# multi_line_output
# 3 = "Vertical Hanging Indent", options here: https://timothycrosley.github.io/isort/#multi-line-output-modes
multi_line_output = 3
include_trailing_comma = true
sections = "FUTURE,STDLIB,DJANGO,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
known_django = "django,celery,corsheaders,strawberry,strawberry-graphql-django"
known_first_party = "uplifty"
default_section = "THIRDPARTY"

[tool.poetry]
name = "uplifty"
version = "0.1.0"
description = ""
authors = ["Uplift <interview@uplift.ltd>"]

[tool.poetry.dependencies]
python = ">=3.11,<4"
celery = "*"
django = ">=4 <5"
django-cors-headers = "*"
django-environ = "*"
dj-database-url = "*"
gunicorn = "*"
psycopg2-binary = "*"
pyright = "1.1.304"
strawberry-graphql = {extras = ["debug-server"], version = "*"}
strawberry-graphql-django = "*"
uvicorn = "*"

[tool.poetry.dev-dependencies]
autoflake = "*"
bandit = "*"
black = "*"
django-dotenv = "*"
factory-boy = "*"
faker = "*"
flake8 = "*"
flake8-datetimez = "*"
flake8-docstrings = "*"
ipython = "*"
isort = "*"
mccabe = "*"
pytest = "*"
pytest-asyncio = "*"
pytest-cov = "*"
pytest-deadfixtures = "*"
pytest-django = "*"
pytest-factoryboy = "*"
pytest-flakes = "*"
pytest-mock = "*"
pytest-pep8 = "*"
pytest-xdist = "*"
safety = "*"
